predictor/requirements.txt
kafka-python>=2.0.2
numpy
pandas
matplotlib
sklearn
joblib